Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -57,8 +57,12 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single('image'), (req
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'update successful' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'update successful' });
+    } else {
+      res.status(401).json({ message: 'not authorizied' });
+    }
   })
 })
 
@@ -96,9 +100,13 @@ router.get('/:id', (req, res, next) => {
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      res.status(200).json({ message: 'post deleted!' });
+      if (result.n > 0) {
+        res.status(200).json({ message: 'update successful' });
+      } else {
+        res.status(401).json({ message: 'not authorizied' });
+      }
     })
 })
